refactor: renames Text component files
