+++ src/components/Text/Text.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { Text } from "./";
+import { Text } from ".";
 
 const meta = {
 	title: "Base Components/Components/Text",
